refactor: index export 추가 및 View 네이밍 변경(View => MovieView)
